Deleted extra labor contents file.
Fixed authors in all Q2 apps.

Changed black to grey in employment, please take a look, but it is tied to the border colors for each of the municipalities. so the map looks different.  Left black to grey in the others the same.

Could we change the legends to say data not available = black, but it wouldn't look very good.

labor plot is still not working, but will fix that.

AR-3-12-15